"Refactor CustomerServiceImpl and update tests"
This commit enhances the `CustomerServiceImpl` by using `CustomerDto` instead of `Customer` and incorporates more checks and validations. The interfaces `CustomerService` and `MenuService` are also modified accordingly. TDD approach is adopted during the entire development process. New integrated tests for `AdminService` and `MenuService` are created while the previous unit tests are kept intact.

+++ src/main/java/com/widetech/menuapp/service/impl/CustomerServiceImpl.java
@@ -1,8 +1,13 @@
 package com.widetech.menuapp.service.impl;
 
+import com.widetech.menuapp.constants.ErrorCode;
 import com.widetech.menuapp.dao.entity.Customer;
 import com.widetech.menuapp.dao.repository.CustomerRepository;
+import com.widetech.menuapp.dto.requests.CustomerDto;
+import com.widetech.menuapp.exception.BusinessException;
 import com.widetech.menuapp.service.CustomerService;
+import jakarta.transaction.Transactional;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,29 +26,77 @@ public class CustomerServiceImpl implements CustomerService {
     private CustomerRepository customerRepository;
 
     @Override
-    public Customer createCustomer(Customer customer) {
-        //if name empty,set phone number as name
-        if (customer.getName() == null || customer.getName().isEmpty()) {
-            customer.setName(customer.getPhoneNumber());
+    @Transactional
+    public CustomerDto createCustomer(CustomerDto customerDto) {
+
+        if (null != customerRepository.findByPhoneNumber(customerDto.getPhoneNumber())) {
+            throw new BusinessException(ErrorCode.PHONE_ALREADY_REGISTERED);
         }
-        return customerRepository.save(customer);
+
+        Customer customer = new Customer();
+
+        //if name empty,set phone number as name
+        if (customerDto.getName() == null || customerDto.getName().isEmpty()) {
+            customer.setName(customerDto.getPhoneNumber());
+        } else {
+            customer.setName(customerDto.getName());
+        }
+
+        customer.setPhoneNumber(customerDto.getPhoneNumber());
+
+        Customer createdCustomer = customerRepository.save(customer);
+
+        return CustomerDto.builder().
+                name(createdCustomer.getName()).
+                phoneNumber(createdCustomer.getPhoneNumber()).
+                build();
     }
 
+    /**
+     * update name, not the phoneNumber, phoneNumber must not be changed once created
+     *
+     * @param phoneNumber
+     * @param customerDto
+     * @return
+     */
     @Override
-    public Customer updateCustomer(String phoneNumber, Customer customer) {
+    @Transactional
+    public CustomerDto updateCustomer(String phoneNumber, CustomerDto customerDto) {
         Customer existingCustomer = customerRepository.findByPhoneNumber(phoneNumber);
-        BeanUtils.copyProperties(customer, existingCustomer, "phoneNumber");
-        return customerRepository.saveAndFlush(existingCustomer);
+        BeanUtils.copyProperties(customerDto, existingCustomer, "phoneNumber");
+        Customer updatedCustomer = customerRepository.saveAndFlush(existingCustomer);
+        return CustomerDto.builder().
+                name(updatedCustomer.getName()).
+                phoneNumber(updatedCustomer.getPhoneNumber()).
+                build();
     }
 
     @Override
+    @Transactional
     public void deleteCustomer(String phoneNumber) {
+        if (phoneNumber == null || StringUtils.isBlank(phoneNumber)) {
+            throw new BusinessException(ErrorCode.PARAM_ERROR);
+        }
         customerRepository.deleteByPhoneNumber(phoneNumber);
     }
 
     @Override
-    public Optional<Customer> getCustomerByPhoneNumber(String phoneNumber) {
-        return Optional.ofNullable(customerRepository.findByPhoneNumber(phoneNumber));
+    public Optional<CustomerDto> getCustomerByPhoneNumber(String phoneNumber) {
+        if (phoneNumber == null || StringUtils.isBlank(phoneNumber)) {
+            throw new BusinessException(ErrorCode.PARAM_ERROR);
+        }
+        Customer customer = customerRepository.findByPhoneNumber(phoneNumber);
+
+        // If customer is null, return an empty Optional
+        if (customer == null) {
+            return Optional.empty();
+        }
+
+        return Optional.of(
+                CustomerDto.builder()
+                        .name(customer.getName())
+                        .phoneNumber(customer.getPhoneNumber())
+                        .build());
     }
 
 }
